+++ src/core/modules/device/device.js
@@ -39,6 +39,9 @@ export default {
       if (device.cordova || device.phonegap) {
         classNames.push('device-cordova');
       }
+      if (device.capacitor) {
+        classNames.push('device-capacitor');
+      }
 
       // Add html classes
       classNames.forEach((className) => {