+++ src/core/modules/router/router-class.js
@@ -1149,7 +1149,7 @@ class Router extends Framework7Class {
     const { browserHistory, browserHistoryOnLoad, browserHistorySeparator } = router.params;
     let { browserHistoryRoot } = router.params;
     if (
-      window.cordova &&
+      (window.cordova || (window.Capacitor && window.Capacitor.isNative)) &&
       browserHistory &&
       !browserHistorySeparator &&
       !browserHistoryRoot &&