+++ src/core/modules/service-worker/service-worker.js
@@ -86,7 +86,7 @@ export default {
       const window = getWindow();
       if (!('serviceWorker' in window.navigator)) return;
       const app = this;
-      if (app.device.cordova) return;
+      if (app.device.cordova || (window.Capacitor && window.Capacitor.isNative)) return;
       if (!app.serviceWorker.container) return;
       const paths = app.params.serviceWorker.path;
       const scope = app.params.serviceWorker.scope;